+++ src/test/java/com/marginallyclever/convenience/helpers/MathHelperTest.java
@@ -1,6 +1,5 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
-import com.marginallyclever.convenience.helpers.MathHelper;
 import com.marginallyclever.convenience.log.Log;
 import org.junit.After;
 import org.junit.Before;
