+++ src/test/java/com/marginallyclever/convenience/helpers/PathHelperTest.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
 import com.marginallyclever.convenience.helpers.PathHelper;
 import org.junit.jupiter.api.Assertions;
