+++ src/test/java/com/marginallyclever/convenience/helpers/StringHelperTest.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
 import com.marginallyclever.convenience.helpers.StringHelper;
 import org.junit.jupiter.api.Test;
